+++ src/models/request.ts
@@ -2,6 +2,7 @@ import { StrUtil } from '../common/utils/str.util';
 import { HttpMethod } from '../common/enums/http-method.enum';
 import { HttpBody } from '../common/types/http.types';
 import { RequestOptions } from './request-options';
+import { ResponseValidator } from '../contracts/response-validator.contract';
 
 /**
  * Represents an immutable HTTP request.
@@ -40,6 +41,7 @@ export class Request {
     public readonly queryParams: Record<string, string> = {},
     public readonly body: HttpBody | null = null,
     public readonly options: RequestOptions = new RequestOptions(),
+    public readonly validators: ResponseValidator[] = [],
   ) {
     this.systemCorrelationId = StrUtil.generateUuid();
     this.timestamp = new Date();